[INTERNAL][FIX] f.DynamicPage get contentWrapper height with clientHeight
BCP: 1780002731

In FLP DynamicPage is tranformed with CSS when user clicks on profile section.
In this Case DP makes wrong calculations because _needsVerticalScrollBar returns result based on scrollHeight and getBoundingClientRect().height.
The value of scrollHeight is returned without the CSS tranformation apllyed but getBoundingClientRect().height returns the value WITH the tranformation applied.
Thats why I refactor it with clientHeight. Unfortunately scrollHeight and clientHeight return different values on IE.
This is reproducible in analytical table scenario. If int he Header there are grid with two items in the cell is reproducible.
Also if overflow-x: hidden; overflow-y: auto;  are deleted from contentWrapper the difference do not appear.

Change-Id: I6d6ebc5426518a543f733e7f32450a7e9b661aca