[FIX] sap.ui.table.Table: improper busy indicator & noData issue fixed
recent changes cause issues with busy handling (when no records are
returned by the backend) and noData state (when a single filter request
is sent which filters the data). This change is a quick fix to resolve
those issues.
A better handling for both of these features can be done in a future
commit!

BCP: 1770044976
BCP: 1770041261
Change-Id: I56bb665bd45ad88c08746dee83c3d88b1c7209bb