[FIX] Configuration: getSAPParam("sap-language") after language change
If the SAPLogonLanguage is set in the configuration with
"setLanguage", the value of "sap-language" in the sapparams map is
now updated correspondingly. The value could differ if another
language is provided as URL parameter. This affected the "sap-language"
GET parameter which is appended to the manifest.json and OData metadata
URLs.

Change-Id: Ia771fdb8dc0ab015b90ceed89d9015d72a9c76c8
BCP: 1770004448
BCP: 1780101258