[FIX] sap.ui.model.analytics.AnalyticalBinding
We set the bNeedsUpdate flag to "true" if ALL request details are empty.
This happens when the initial analyticalInfo is empty and is not set correctly
before the control calls getRootContexts etc.
If at a later point the analyticalInfo is correctly set AND the bNeedsUpdate flag
is still true we falsly force a change event in checkUpdate
--> this might lead to an unnecessary re-rendering of the control.

In case we have at least 1 valid request (including measures and/or dimensions)
we set the bNeedsUpdate flag to false, because the update flag is set to true
ANYWAY during the response-processing.

Change-Id: I7ffd144f03c61610bdabe3dd6cfc1ffd06750a0e
BCP: 1770008178