[INTERNAL][FIX] ESLint: Fix linting via "grunt lint"
With faffe40664a4cb7e83e605a15f2e4c4ff493e0ce some entries in the
.eslintignore have been introduced which broke the whole validation via
"grunt lint". The change looks valid so this might be a bug in either
eslint or grunt-eslint. As we are using outdated versions it doesn't
make sense to have a deeper look into this.

This change removes the re-includes as the code
contains errors which need to be resolved beforehand.

Change-Id: I0ff8e7d24fb3b7880a60420d903900790d7aa7a8