[INTERNAL] Internal switch to CalendarDate for calendar based controls.
There is not built-in JavaScript object/class for just a date. It is only the JavaScript Date object (for short jsdate)
that keeps both date and time information within. Although a date is time and timezone irrelevant (see
link https://en.wikipedia.org/wiki/Calendar_date)

Due to this every time where pure date calculations are needed, the jsdate object, which is usually presented
 in a local timezone is being represented with the same datetime values, but considered as UTC ones.
The idea is to have one object representing dates in order to will simplify the implementation and reduce the
 errors during conversion.

Implementation hints:
- every date passed to an internal method of a behavior class or any method of a renderer is considered as a CalendarDate

- almost all usages of CalendarUtils._createUniversalUTCDate and CalendarUtils._createLocalDate are removed. Alternatively
 CalendarDate.fromLocalJSDate and CalendarDate.prototype.toLocalJSDate are used

- the CalendarDate is used internally only. No standalone usage.

- introduce new utility that works with CalendarDate - CalendarUtils._DATE (for internal use only). The utility compares
 and calculates dates.

- all  implementation files except PlanningCalendar are switched to CalendarDate. The use of UniversalDate is removed.

- PlanningCalendar, DatePicker, CalendarTimeInterval are not switched, as they maintain info for real datetimes. Still, they are modified,
due to the cases where interact interacts with some of the rest (already switched) by use of a private/protected methods.

- Public used date properties and API methods are not changed.

- Classes that are swtihed to the CalendarDate have such comment at the top.

- Classes that are sitched to the CalendarDate remove direct usasge of UniversalDate.

Change-Id: I453e70926220fd97eb89b05f9539abddcd873522