[INTERNAL][FIX]sap.m.Image: Ensured cached image added to DOM
This fixes the follwoing problem for sap.m.Image with setup:
- iOS mobile device platform
- image mode is sap.m.ImageMode.Background
- the resource at the given image URL is *cached*

The problem was that the "background-image" css was *not added* to the DOM element of the image,
because the "background-image" css was added to the old (or invalid) jQuery DOM reference [during rendering],
BEFORE the new (and valid) DOM element was flushed/actualized by the RenderManager.

Note: The problem does not reproduce on desktop(/desktop emulator),
because on desktop a load event is fired even for cached images and the problem is surpassed there.

Change-Id: Ief2b13daba019de858a8bf6e6709018cc5d094d5
BCP: 1780106579
(cherry picked from commit f69ed4824f64d11d58854852ccb30f935061bc4f)
CR-Id: 002075125900001137312017