[INTERNAL][FIX] sap.m.ObjectStatus: Height when used in sap.ui.table.*
Fixes issues introduced with change:
Ieb81bbfde3f0681efdfd25be8ed3a5c50c2e4991

When used as a column template in a sap.ui.table table, the actual
row height could become bigger than the default row height in
condensed density. This causes vertical scrolling issues when the
visibleRowCountMode is set to "Auto".
Furthermore the text is again properly truncated with an ellipsis,
when it does not fit into the column.

Change-Id: I8f179adc051df7238fe4d0a92e215f01aa8f7390