[FEATURE] sap.m.TimePicker: added startValue Property
The TimePicker element defaults to the current time when the
popup is opened and no initial value is set. There is currently
no options to override this behavior.

The TimePicker now has an additional property to allow developers
to pick a default value to be displayed in the slider popup.
- When no value is set, it defaults to the current time (default
  behavior).
- When a value is specified, it always defaults to the startValue
  property

The onBeforeOpen handles checking this value prior to displaying
the sliders.

Fixes https://github.com/SAP/openui5/issues/1312