[FIX] Component Lifecycle: ensure to load includes during Metadata#init
1.) ensure that the Metadata#init is executed before the
ComponentMetadata#onInitComponent. The dependencies are now loaded again
before the includes (JS and CSS). This restores the loading behavior to
the execution order before the Component Lifecycle change.

2.) Manifest loads the includes in the onInit again to ensure
compatibility and introduces a check for the load and removal of the
includes to ensure the loading or removal being done only once.

Change-Id: Ifd7d6de2b47cb9397fddd4d378a203382461c63c