[INTERNAL] Component: JSDoc fix for the createComponent function
The JSDoc plugin doesn't detect a colon as ending of a line for the
summary and looks for the next available dot which is part of the <pre>
section in the JSDoc for the createComponent function. This creates a
strange JSDoc HTML document. This change fixes the issue by explicitly
having a summary line ended with a dot.

Change-Id: I82b20ae5cc5ad7b57840244697695e1f891024b9