[INTERNAL] Table: Improve RefreshRows
When the rows aggregation binding is refreshed, new contexts are
retrieved. Subsequently the binding length is updated and in the same
step some UI parts are updated. As the binding length is always 0 on
a refresh, the table always scrolled to the top and issued a second
request.
When the rows aggregation is refreshed there should not happen any UI
 updates.

Change-Id: I58769e1d4f37a9ec148cb00c5128d4165d6335f7