[FIX] v2.ODataListBinding: Remove pretend request
In the old ODataModel we pretended to send a request when the $count
request returned 0 and no data request was sent, to have events fired
for controls as if a request happened. In the v2.ODataModel the $count
request is no longer synchronous, but sent asynchronously together with
the data request, so the mechanism is no longer needed/valid/useful.

BCP: 1770146794
Change-Id: I6c6b3a58ba9843783734eefb140762509d23cc71
CR-Id: 002075125800001669522017