[INTERNAL] sap.m.FacetFilter - update doc to reflect to real implementation.
According to the implementation the "liveSearch" property controls the search
behavior in FacetFilterLists. The doc is saying the opposite and is corrected.

No other search fields are subject of controlling the "liveSeach" behavior.
The search field that appears when a dialog is opened (i.e. not in the details of a list),
is in "liveSearch" mode permanently.

BCP: 1770185925

Change-Id: If1123fdd604dc7c0dbbed5797c4f5a1ee0ff758c