[INTERNAL][FIX] FakeLRepConnector on FLP
The current FakeLrepConnector has 2 problems when running in real FLP
- In FLP there are many app component can be loaded in one browser
session. When FakeLrepConnector is enable once, it impacts all follow
app components if disable function is not called.
- It is possible that the ChangePersistance load changes before the
connector is mocked. Hence, the new change will be sent to the real back
end.
Therefore, to enable FakeLRepConnector on FLP, appComponentName and
appVersion should be provided to mock connector of a specific component.

BCP: 1770178900

Change-Id: Ic14f5f27b89d1cda4730dfcd7b6b47e988a81124