[FIX] sap.m.PlanningCalendar: appointments are correctly positioned when property showRowHeader is set to false
If showRowHeader is set to false, the header(1st column in the table is not shown).
This may cause a problem with sap.m.Table when the second column is a popin one, where once column's minScreenWidth
 is reached (i.e. less than defined) the 1st (invisible) column occupies its original space and shifts the second to the right.

The sap.m.Table relies that if there is a popin column, there will be always at least one visible other column. This is according to
the spec and may be changed with a feature request.
This change is a temporary legal workaround of the issue unless it is resolved by the sap.m.Table.

(manually cherry picked from Idb7b11063ff7896fd8b26ef0724383f041f57f8e)

BCP: 1770208038

Change-Id: If0415b2aed19c911b991c5116864b378541567de
CR-Id: 002075125800001674212017