[FIX] sap.m.RangeSlider: Change & LiveChange events are now fired properly on click.
When the user clicks on the RangeSlider the handles move,
values are updated, but the change and liveChange events are not fired.
This was because the range is already updated (by this._updateHandle)
and the check for range modification in _ontouchend method would fail.
So, after we're sure that there's a change in the range, we need to
 trigger these events manually.

BCP: 1770103593

Change-Id: I19b530c0658aa876a3d44d4b732279b7a932a182
CR-Id: 002075125900001431022017