[INTERNAL] sap.ui.demokit: hide controls again from public documentation
All controls in 'sap.ui.demokit' are non-public controls. Their usage
always was and still is restricted to the SDK and to legacy SDK content
developed by SAP (e.g. interactive docu pages for the now-deprecated
sap.ui.commons and sap.ui.ux3 libraries).

In the old API reference, the above restriction was communicated by
completely excluding 'sap.ui.demokit' from the API reference at build
time. The new API reference displays content for all libraries found in
sap.ui.getVersionInfo(). For the sake of completeness, 'sap.ui.demokit'
must be part o that list and therefore its controls mistakenly appear as
'public' in the new API reference.

As a fix, the visibility of the 'sap.ui.demokit' controls and types has
been adapted to @sap-restricted and the @ui5-metamodel tag has been
removed. This now consistently reflects the above usage restriction in
the source code as well as in the old and new API references.

Change-Id: Id2204e0f7bdefd1470039689fb20cc2a1e711d9f