[INTERNAL] odata.v4.*: PATCH when Context does not point to the entity
Now the property binding does not rely on its context to point to the
entity, but searches it in the path. So the binding's relative path may
also contain a navigation property, or the context may point to a
ComplexType.

(cherry picked from commit 2b6d52febe1867b2658675beefde6965db553327)

Change-Id: I2de6894697fcc1516a4399c458cd5d4a48384cd1
JIRA: CPOUI5UISERVICESV3-434
CR-Id: 002075125900001907642017