[INTERNAL][FIX] sap.m.IconTabHeader: Ensured model propagation into popover
On press of the overflowButton of the IconTabHeader, the internal selectList is set as child of the internal popover.
To ensure propagation of bound properties to the popover children (e.g. the selectList), this change adds the popover to the "dependents" aggregation of the IconTabHeader.
JIRA: BGSOFUIPIRIN-1588

Change-Id: Ib5ef5520f6c09162ed70bef6f7ad501c2669c6c3
(cherry picked from commit d39644061c1a87366f903d4e824417ccef8d5d4e)
CR-Id: 002075125900001691922017