[FIX] v2.ODataListBinding: Server operation for expanded collections
With the introduction of clientside sorting/filtering in the
ODataListBinding we did activate client operations automatically for
expanded lists, where we know all entries are already available on the
client. To enable applications to explicitely use server side operations
on expanded collections, this change introduces a new operation mode
"Default" which keeps the current behaviour and changes operation mode
"Server" to force server side operations also for list bindings on
expanded collections.

BCP: 1770152832
Change-Id: I3faf9e661a90b0ab40ceabc14034ee4795e31096