[FIX] handlebars: ensure proper context stacking in compat mode
handlebars 4.x introduced an optimization for the context stacking which
avoids stacking the same context again. In case of "compat" mode and
relative parent context access this breaks the relative structure and
the context value lookup points to the wrong context. In case of
"compat" mode the context stacking must be properly ensured!

Change-Id: I6bdd249186984db8a3b04f83f8d59af073be7a85
BCP: 1770235263
CR-Id: 002075125900001776572017