[INTERNAL] sap.ui.layout.BlockLayout: Switch cell color properties to enumeration.
backgroundColorSet and backgroundColorIndex were defined as integers
which caused many discussions and confusions.
Now these properties use enumerations and there's validation out of the
box. Also now it's clear that we cannot define limitless color sets for
cells' background color.
Background coloring is moved to the CSS.

BCP: n/a

Change-Id: If6d0761388cd2aeb309ef99d5fc22d7bce9ba0b8
CR-Id: 002075125900002149492017