[FIX] sap.m.DateTimeInput: value property binding type different than control type is supported now.
When the 'value' property is bound with given type (binding type) and the pattern of the binding
type implies difference with the DateTimeInput 'type' property, the binding type takes precedence.
In this case the 'type' property does not change, but the underlying picker implementation switches
according to the binding type.

For example, if the type is sap.m.DateTimeInputType.DateTime, but the binding type is sap.ui.model.type.Date
the used picker is sap.m.DatePicker, instead of sap.m.DateTimePicker.
With the previous implementation, this scenario fails at the DateTimePicker trying to open the time sliders.

Change-Id: Ie9b5620a1164932bd70c580b4896471d8a1b1376
BCP: 1770136207