[FIX] v2.ODataListBinding: Check for expanded data in checkUpdate
When a relative list binding was created without expanded data available
and a followup request of the context entity did contain expanded data
for the list, the expanded data was not updated for create/delete
operations. To address this in method checkUpdate we now also call
checkExpandedList to properly detect expanded list data.

BCP: 1780081156
Change-Id: Ic9a6c6f5d493935084d7519ad95bf6d2a3e2eaa3
CR-Id: 002075125900003345672017