[FIX] sap.m.DatePicker: Choosing date equal to the mindate/maxdate from picker is supported.
When the min/max date is set with a value (JSDate) that includes time part and the user picks a date
that is on the same calendar date(date of year) as the min/max, this is not possible.

The reason for this not working before is that the picked date(i.e. 01.01.2017 00:00:00) is considered before the minDate (e.g. 01.01.2017 13:00:00), if time part to the min date is set.
Fixes: https://github.com/SAP/openui5/issues/1461
Change-Id: Ib71d0afc0bfc6349c000b0057768fb1278f3d229
BCP: 1770203308