[FIX] core.IconPool: Use local SAP-icons font as fallback only
With 31ac6227f457e768a1077b6de2a614c8c055175a an issue regarding blocked
fonts has been resolved by also allowing a locally installed SAP-icons
font to be used.

However this caused issues like missing / old icons in case
the font has been installed localled but doesn't fit to the UI5 version
in use.
This issue is quite hard to spot and would require to always align the
SAP-icons font installed locally with the font used within UI5.

Therefore the order of referencing the font src is changed to only use
the local font as a fallback.
This is also the exact use case of the issue solved by the mentioned
commit.

Change-Id: Idaa1bbf4e0152e767d9426d2cfa7294403bf9acc
Fixes: https://github.com/SAP/openui5/issues/1508