[FIX] sap.m.DateTimePicker: Open Popover in small screens corrected
Now in Chrome when the size of the window is really small, when open the 
popover it stays opened.
Also there's a screen reader improvement coming from this change.
Before it, when open the popover the screen reader red the
selected/focused date, month, year, week number and day of the week +
everything on the screen because screen readers find the popups like a
window and they read everything. Now it only reads the selected/focused
date, month, year, week number and day of the week.

Change-Id: Ie3b4e0547e25df4345d13a3eaf2b953f9e370650
BCP: 1770203362
BCP: 002075129400006237152017
CR-Id: 002075125800005029832017