[INTERNAL] fl can request changes from a passed url
The flexibility targets via a hard coded url pattern the
"/sap/bc/lrep/flex/data/<variable parameter>" route. In addition the
variable parameters are determined at run time even if the result is
already present in the asyncHint.
This change enables the corresponding classes to pass the url form the
async hint until it reaches the LRep connector. The connector does not
make the calculation in case the url was passed and sends the request
based on the passed url.

Change-Id: I8555f2dda826e1cdf437837e70aaf543ff186534