[INTERNAL] fl determines its asyncHints more precise
Previous the asyncHints used on a component creation were only
identified by the first appearance of the "sap.ui.fl.changes" within the
name of the asyncHint. Afterwards the reference was checked and in case
the reference didn't match no asnycHint was used. This is a valid
behaviour in the current state but if multiple asyncHints with the name
are present (possible in the future) this mechanic will not be
sufficient. Therefore the reference comparison was  moved to the search
itself.

Change-Id: If0e584a80cc44922d82784857adc85ebf124d539