[INTERNAL][FIX] sap.ui.documentation: Corrected scrolling in samples master list
The scrollEnablement of sap.m.Page is fully initialized only after the onAfterRendering delegates of sap.m.Page are called (scrollEnablement initializes in its own delegate to onAfterRendering of sap.m.Page).
This is why calls to scrolling methods on the page should be queued after onAfterRendering of the page.
Snippix: 76337

Change-Id: Ia09d06fa5f3ed9d5aa8f1d159a374e3d63ce5ca6