[INTERNAL] sap.m.Button: Split/combine custom handler update
After this update, the buttons that are to be combined are
stored as dependents of each MenuItem that is being created
in the common MenuButton. The id of each button is also stored
so that upon split, the corresponding buttons can be taken
out of the MenuItems and inserted in their previous state.

Change-Id: Ic1db52c89619093868d8e7befe9d842eaf5ac273
CR-Id: 002075125800002470712017