[FIX][INTERNAL] fl XMLPreprocessor detects sap-app-id
BCP: 1780217772

The sap-app-id within the application startup parameter was used to
detect a different set of flexibility changes. For compatibility reasons
this determination is still in the code, but was removed with Change-Id:
I707885766435a38b7d3d31dd58f0e9d6b7d03293. This changes adds an
additional test for this parameter and corrects the behavior.
 
Change-Id: I3d0313f5f98d8c895d242ef337b5d13f14ca3381