[INTERNAL] FragmentControl: Encapsulate creation of MOM
This is done in order to be able for own Fragment Controls to define
their own MOM via _getManagedObjectModel.

The idea behind that is that e.g. fragment control like the filter bar
might want to have a richer MOM then just the normal MOM, in the POC I
currently work on this is an MOM that is also a ConditionModel.

Change-Id: I3827ebe405e08969c3b627617ebbc8986cec2f29