[INTERNAL][FIX] core.Component: register URL early
Partial revert of fd40a1df. When a URL is given, it has to be registered
even before a manifest completed loading asynchronously as the Manifest
in its constructor already determines the base URL for the component.

This revert needs to be revised as it again forces a "manifirst" caller
to know the final name of the to-be-loaded component in advance, before
the corresponding manifest has been loaded. 

Change-Id: I2e62cf6775e0b09b051a71bf3e7b56da04de91a8
BCP: 1780249208