[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Corrected scrolling to section
When the page is scrolled with 0 delay, the function that toggles the anchorBar position (removes/adds the anchorBar from/to the scrollable area ) is called *after* the scroll completed.
Therefore the anchorBar was removed from the scrollable area *after* the page arrived at correct scroll position => extra pixels were scrolled (with y === anchorBar height).
Fixed by ensuring the header is toggled before the scroll completed.

Change-Id: I561fcfd178e84af6ff38715316f4850495b7688f
BCP: 1780236313