[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Corrected scrolling to section
When the page is scrolled with 0 delay, the function that toggles the anchorBar position (removes/adds the anchorBar from/to the scrollable area ) is called *after* the scroll completed.
Therefore the anchorBar was removed from the scrollable area *after* the page arrived at correct scroll position => extra pixels were scrolled (with y === anchorBar height).
Fixed by ensuring the header is toggled before the scroll completed.

Change-Id: I3ee07f806fdeba348289c30b3855813884de8533
BCP: 1780236313
CR-Id: 002075125800002962862017