[INTERNAL] sap.m.TimePicker: Placeholder is now properly added to accessibility info
TimePicker inherits two artifacts that hold useful information which can be used as a WAI-ARIA label or description - placeholder and tooltip.

To do this, we need to render a hidden span element with a value that is the same with the value of the corresponding artifact (if present) and associate it by adding its ID to one of its 'aria-labelledby' or 'aria-describedby' DOM attributes.
We also have into consideration that its 'displayFormat' pattern is assigned as a default placeholder if no placeholder value is set by the application developer.

In addition, it can also be referenced with external label like any other Input control either by using the 'labelFor' association of the given label or by using its own 'ariaLabelledBy' association.
By design, it has an ARIA custom string role which is also rendered as a hidden child span element with particular ID and added to its 'aria-describedby' DOM attribute.

BCP: 002075129400002214112017
Change-Id: Iffff3cd21ae49ed3fbe51b8ee6e9fafe68c9f8ac