[INTERNAL] mvc/XMLView: no error when cache key is missing
Before this change the XMLView made use of jQuery.sap.log.error if a cache
key was not delivered by the according Promise. That made it impossible
for adopters to handle a missing cache key increment silently or in their
own manner. Hence we do only log a debug message now in case a cache key
could not be provided.

Change-Id: I363cb35ab46257392779610ccb26023897df079a