[INTERNAL] New demokit - explored index no longer redirects for theme designer
The explored.html file now checks whether it has been loaded in an iframe by theme designer.
If so, it does not redirect, but rather bootstraps the app directly, just as index.html would.
In addition, it sets the hash to "#controls" so that the explored section is selected by default.
Otherwise (not in an iframe or not loaded by theme designer), it redirects to index.html.

Change-Id: I86f4dc00ccaf05e41858a118184df7b2fa303fd6