[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Corrected restore scrollPosition
Restore scroll position didn't check if the latest saved scroll position is still valid.
However, the saved position will be invalidated if a section preceeding the selected section is no longer available (because of being deleted of hidden).
Fixed by adding a check that the restored scrollPosition is within the section that should be selected.

Change-Id: I53d45e65d0cef536b57321b2f42f18815c5d9032
BCP: 1780236313