[INTERNAL] sap.ui.documentation - Documentation tree restructured
The content of the main tree node now serves as top-level content of the tree.
Test pages are still preserved, if available (as before).
Additionally, a bug was fixed by clearing the selection of the tree when no topic is selected,
allowing the user to successfully open the first entry upon returning to the documentation section.

Change-Id: Idf47f6625605db4aefe09edc9e0efa8d3a41f0b3