[INTERNAL][FIX] sap.ui.fl.Change - Validation of change's state
- Since the DIRTY state of a change is no longer determine by the change
content but explicitly set in the setContent function. After creating a
change from change constructor and if setContent function is called, the
change has a incorrect DIRTY state. So that prevents of saving change to
the back end. This change implement a validation of state before
setting, which does not allow to move directly from state NEW to DIRTY.

Change-Id: I57938ca1b68fbf1f25d700613742ee9198f9fa01
BCP: 1780313680