[INTERNAL] sap.m.Popover: Close behavior Explored sample added
A sample demonstrating how the close behavior of the sap.m.Popover can be controlled.
When the popover is opened, add a style "pointer-events: none" to the area that has
not to be triggered and when the popover is closed, remove it,
e.g. in order to prevent drill-down navigation.

Change-Id: Ia3660daee863db1d985168e3eb7a975b138ba162