[FEATURE] unified.FileUploader: Allow processing of files before upload using interface
This change introduces the new interface sap.ui.unified.IProcessableBlobs and changes the sap.ui.unified.FileUploader control to use this interface when upload() is called.
It makes it safe to implement custom controls that extend sap.ui.unified.FileUploader and which modify the Files before they get uploaded.
This comes in handy in use cases such as creating zip file for the upload.
Internally Promises are used.

Change-Id: I5ff85b6da716d400826ee1f1f544f05d0e821706