[FEATURE] FileUploader: allow processing of files before upload via interface
This change introduces the new interface sap.ui.unified.IUploadableBlobs and changes the sap.ui.unified.FileUploader control to use this interface when upload() is called. It makes it save to implement custom controls that extend sap.ui.unified.FileUploader an which mobify the Files before they get uploaded. This comes in handy in use cases such as creating zip file for the upload. Internally Promises are used.