[FIX] sap.m.PlanningCalendar - appointment row focus outline covers all 4 sides for High Contrast Black & High Contrast White themes.
IE and edge have an issue with outline when padding of the element is 0 (outline is partially/fully invisible)
The issue is that when there is a next row td with padding-top 0, the previous row's outline is not visible.

The solution actually preserves some space for the row focus outline for all themes. This is aligned with the corresponding visual designer.

(cherry picked from commit a3a5722f891643510e4de089485475dbebd36f5f)

BCP: 1780142233
Change-Id: I8aefca5b744cf0f71fb0f1ce2f77ba76349c2f07
CR-Id: 002075125800003764292017