[INTERNAL] sap.m.MenuButton: Split handler updated
This change resolves issue with finding the index of
control in multiple aggregation, when the parent control
is overwriting its aggregation methods. In this case
calling the default Parent.indexOfAggregation will result
in wrong index. To illustrate it, if we are looking for a
control in the aggregation called "actions" the method
called on the parent will be Parent.IndexOfAction.

Change-Id: Ia347074f400f2b903e9bf66464cedd7697e6d206
CR-Id: 002075125800004024542017
BCP: n/a