[FIX] changes are applied on recreation in initial app startup
In case flexibility changes were not in the application, the fl library
is loaded lazy. This results in no registration of the
propagationListener in to the component initiation BEFORE the
component was created.
If RTA / personalization takes place in this component the changes are
applied ad hoc but are not propagated.
In case a control was destroyed and the change must be applied again,
but this never happened therefore this results in changes being "lost in
this session".

Change-Id: I7ef1928da749259b2f2a9c87db2bf691c5d9daf0