[INTERNAL] ColumnHeader: Implementing requirements for screen reader support and relevant CSS styles adapted to correctly render the control in Horizontal alignment
A hidden span element is rendered within the control if the control is
interactive. This hidden span contains the relevant information such as
the ColumnHeader text, sorted, sort order, filtered and other text
information that describes the control and its state. Also relevant CSS
styles were adapted so that the control is rendered correctly in
Horizontal alignment specified by the Column.

Resolves:
JIRA: CPOUIFTEAMB-881
JIRA: CPOUIFTEAMB-882

Change-Id: I91bcc173998986ccd929d7766d8cd2b50a5f53d3