[FIX] core.ClientTreeBinding: Fix for relative treebinding
In case the sap.ui.model.TreeBindingAdapter is applied to the
treebinding an artificial root node without any context will be added
to the tree. When the tree binding gets bind to a relative path the path
could not be resolved. Hence the magnitude of the root node was not
correctly calculated and the tree not properly created. The magnitude
gets calculated by the _getChildCount function of the
sap.ui.model.ClientTreeBinding.

With this fix the relative path gets resolved as an absolute path by
an intermediate step in the _getChildCount function in order to calculate
the correct magnitude.

Change-Id: I336c4753f12f0d8b472babd8c1baeff812a02ba2
BCP: 0020751294 0000379717 2017
CR-Id: 002075125800003948572017