[FEATURE] Synchronous context creation
OdataContextBindings could create a preliminary
Context, so dependent Bindings could load its data 
in parallel. Two flags are provided to control the
new behaviour:
createPreliminaryContext and usePreliminaryContext.
This flag could be passed as parameters to each
binding. The v2.ODataModel provides a global flag
preliminaryContext to enable the new behaviour for
all bindings.

Change-Id: Ib5779471761f188b897690c8bf00530143e2fcca
BCP: n/a
CR-Id: 002075125900004069832017