[INTERNAL] sap.ui.documentation - Tree improvements
This change addresses two issues with the documentation section's master controller:
 - The model, used for the documentation tree, used to have a size limit of 100.
This has been increased to allow expanding the whole tree.
 - The filtering in documentation was done based on the "text" field of each topic.
In contrast, the search in API Reference was done based on "name", which contains
the whole namespace of the entity, thus giving better search results.
To achieve a similar behavior, a "name" field was introduced for documentation topics too.
This new field consists of the parent node's "text" field, concatenated with the current
node's "text" field. Thus, when a node is matched in the search, all its children are matched too.
Since both ApiMaster and TopicMaster controllers now use the "name" field for filtering,
the parent onTreeFilter method can be reused entirely for both controllers.

Change-Id: I7eca236aeb80897ff1f3ec3309caf4d277629710