[FEATURE] sap.m.IOverflowToolbarContent interface introduced
This change allows any control to define its own behavior when inside sap.m.OverflowToolbar.

Until now, controls needed to be "whitelisted" in order to be able to move to the overflow menu.
This whitelisting consisted of the following:
 - an entry was created in OverflowToolbarAssociativePopoverControls._mSupportedControls
 - a pair of methods _preProcess*ControlClass*/_postProcess*ControlClass* were added again
in the OverflowToolbarAssociativePopoverControls helper class
 - CSS to be applied when the control goes to the overflow menu was added in the file:
base/OverflowToolbarAssociativePopover.less

With the addition of the new interface, each control can provide this data on its own,
without the necessity to be whitelisted in a central helper class. See the description
of the "sap.m.IOverflowToolbarContent" interface in the library file.

Change-Id: I57474c2eb5d0269ebbb7a7e24e9388bfeff17804