[FIX] Table: Correctly process focus events related to the edit mode
Some input controls with an open button (e.g. sap.m.ComboBox) caused
issues with the keyboard handling of the table, if the user clicked on
the open button. The focus events were not correctly processed, which
could result in several focus changes and changes between navigation
and edit mode.

Change-Id: I2807166bc47ac3d6e74ec1b82ab600e57c572fde
BCP: 1780315614
CR-Id: 002075125800004092112017