[FIX] sap.m.ViewSettingsPopover - title is always propagated to dialog aria label.
By default the aria-labelledby attribute of the popover is set to the toolbar that contains the segmented button with sort, filter & group.
With this change, the aria-labelledby refers to an invisible element that is being updated with the current page title (Sort, Filter, FilterDetails, Group),
i.e. the text of the invisible element considers any user navigation.

This element existed before, but was not correctly referenced/used.

Change-Id: Ib9b162158ee6fce146e0fd5a64fc56338200e2e6
BCP: 1780141942