[FIX] sap.m.ViewSettingsPopover - title is always propagated to dialog aria label.
By default the aria-labelledby attribute of the popover is set to the toolbar that contains the segmented button with sort, filter & group.
With this change, the aria-labelledby refers to an invisible element that is being updated with the current page title (Sort, Filter, FilterDetails, Group),
i.e. the text of the invisible element considers any user navigation.

This element existed before, but was not correctly referenced/used.
(cherry picked from commit 0696ddf1fd0d73d5a363bc3f3ffc3451ca4999f1)

Change-Id: Ib9b162158ee6fce146e0fd5a64fc56338200e2e6
BCP: 1780141942
CR-Id: 002075125800004116992017