[INTERNAL] sap.f.FlexibleColumnLayout - invisible columns are also not focusable
BCP: 1780166796
When some FlexibleColumnLayout column is in fullscreen mode, it takes up 100% width
and the other columns get zero width. However, all elements on these zero-width columns
are still focusable as nothing prevents that.
Setting tabindex to the hidden columns does not help as their children remain focusable.
This change makes zero-width columns also with visibility: hidden to make them, and their children, unfocusable.
This visibility change though has to happen after 560ms (after the width animation is over),
therefore it is added to the animation selector (although visibility is not actually animated)
so that it is delayed by exactly this amount of time.

Change-Id: If83293e5718ddc68c3a1a4cd5dce410523318109