[INTERNAL] sap.m.Select - change event no longer triggered excessively
BCP: 1780153332
Changing the value of sap.m.Select programmatically while the control has focus
used to desync the _oSelectionOnFocus variable. This change synchronizes it with
the newly selected item so that once the select loses focus, the change event is not
fired yet again (due to the focused item being different than the selected item).

Change-Id: I61e159274df828d823baa73083a4f2738ffea58f