[FIX] sap.m.MultiInput: Do not tokenize text when ValueHelp is opened.
When there was a valid text in the input and the ValueHelp was triggered,
that text was tokenized as the ValueHelp Dialog is opening.
This is unexpected behavior. Instead the text should be leaved untouched
and when the ValueHelp gets closed the text should be highlighted as it
is in sap.m.Input.

BCP: 1780279195

Change-Id: Ida15b10e94a304456d5c94824a592a4e89ade79d