[INTERNAL] sap.m.NavContainer - navigating back to the same page no longer throws an error
fixes: https://github.com/SAP/openui5/issues/1651
BCP: 1780393005
The method "_safeBackToPage", used in routers, sometimes initiated a back navigation to
the currently selected page (for example when navigating to another route with the same target),
which lead to an error in the console.
This change prevents such a navigation in the first place as it is redundant.

Change-Id: Id3160c3c7b1e8ed413c066ecc68bcda73462fc57