[INTERNAL][FIX] sap.m.StepInput: properly set valueState and valueStateText properties to the control
Now sap.m.StepInput pass valueState and valueStateText properties to the input aggregation, but also set its own properties. In this way the control properties (valueState and valueStateText) are properly changed and there is no need of overriding the getter methods

Someone can call low level API such as getProperty or setProperty which is not recommended but would result in workaround the StepInput implementetaion for get or set of the valueState and valueStateText.

Change-Id: Ia3ab1f8dca963838a22d846b89a295c8c257f297
BCP: 1770331743