[INTERNAL] sap.m.OverflowToolbar - suppress calculations until theme is applied
Background:
During the onAfterRendering phase OverflowToolbar calculates the widths of its children.
This however is not accurate until the theme is applied and the children are styled.
Therefore, OverflowToolbar also listens for the "themeChanged" event and invalidates itself
so that it can repeat the calculations once the theme is applied.
This change makes it so that calculations are not performed altogether until the theme is set.
BCP: 0020751294 0000428939 2017

Change-Id: I3742016ad83ce17c82ef3c042bcdb91f5200bfd9
(cherry picked from commit 20937e4b68c319bd8af2190b74724dce4db9eb44)
CR-Id: 002075125900002667172017