[INTERNAL][FIX] sap.m.Page: handling of showFooter on first rendering corrected
The class '.sapMPageWithFooter' is added, when a footer exists,
but the 'showFooter' property`s value should be taken in account as well.

The class adding/removing is handled in the 'showFooter' setter,
but it would work runtime, when the DOM is already in place
and would not work on first rendering.

Changes:
(1) 'sapMPageWithFooter' is added, if a footer exists and 'showFooter' is true
(2) unit tests are added to cover the scenario

BCP: 1770343643
Change-Id: I77c2a29f8b0c0e6fb3110a5acaa60cd611d74fdd