[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Corrected position of top section
The scroll position of the first section should be 0 (top of the scrollable container) in expanded mode.

The old calculation unnecessarily depended on assumptions of the structure of the page (e.g. presence of anchorBar) which are not always valid.

Change-Id: I6b1dff37567410e1b6a9aeb265f402dc18722f26
BCP: 1780365635
(cherry picked from commit b9f2b4001244622be33722c96e86ae8b3414ecab)
CR-Id: 002075125900002794022017