[INTERNAL] sap.ui.core: Invalidation handling during rendering.
 - If the parent of the control is invisible or never rendered yet it
should not be necessary to bubble up the invalidation to the parent
since we need to wait the next rendering tick of the parent  anyway to
see the rendering results.
 - If the parent does not have bOutput property, it is probably an
Element or UIArea then we should let the invalidation bubble up to the
parents e.g. for the initial rendering.

 - This change will shorten the invalidation hierarchy when the parent
control is invisible or not yet rendered thus we will rerender the
control hierarchy less and initial rendering will be faster.
 - With this change at the renderer/onBeforeRendering, control
developers can call the public mutator method of the control freely,
instead of calling the protected version of the mutators(setProperty,
addAggregation...) with suppressing the invalidation because now the
parent invalidation during rendering will be ignored and wait the next
rendering tick of the parent.

Change-Id: If31ad6fa79c589c2f613f781a321267058306903